feat: 添加仪表图组件
