+++ apps/web/src/adapter/reducers/getAggregateHistory.ts
@@ -1,22 +1,47 @@
+import { EntityAPISchema } from '@/services/http';
 import type { PluginProps } from '../types';
 
-export interface ChartShowDataProps {
-    entityLabel: string;
-    entityValues: (string | number | null)[];
-}
-
+type AggregateHistoryResult = EntityAPISchema['getAggregateHistory']['response'];
 export const useReducer = () => {
     const run = (
-        historyData: (SearchResponseType<EntityHistoryData[]> | void)[],
+        aggregateHistoryData: (AggregateHistoryResult | void)[],
         viewProps: PluginProps,
     ) => {
         const { config } = viewProps;
         const { entity } = config || {};
 
         const entityList = Array.isArray(entity) ? entity : [entity].filter(Boolean);
+
+        const newChartDatasets: AdapterResult[] = aggregateHistoryData.map((historyData, index) => {
+            const { value, value_type: valueType, count_result: countResult } = historyData || {};
+
+            if (countResult) {
+                return {
+                    entity: entityList[index],
+                    data: (countResult || []).map(item => {
+                        return {
+                            value: item.value,
+                            valueType: item.value_type,
+                            key: item.count,
+                        };
+                    }),
+                };
+            }
+
+            return {
+                entity: entityList[index],
+                data: [
+                    {
+                        value,
+                        valueType: valueType || null,
+                        key: '',
+                    },
+                ],
+            };
+        });
+
         return {
-            chartLabels: (entityList || []).map(item => item?.label),
-            chartDatasets: (historyData || []).map(item => item?.value),
+            chartDatasets: newChartDatasets,
         };
     };
