--- apps/web/src/plugin/plugins/gauge-chart/view/hooks/index.tsx
@@ -1,1 +1,0 @@
-export { useSource } from './useSource';
