--- apps/web/src/plugin/plugins/gauge-chart/view/style.less
@@ -1,27 +1,0 @@
-.ms-gauge-chart {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    width: 100%;
-    min-width: 100%;
-    height: 100%;
-    padding: @padding-md;
-    background: var(--main-background);
-
-    &__header {
-        font-weight: @font-weight-bold;
-    }
-
-    &__content {
-        display: flex;
-        flex: 1;
-        align-items: center;
-        justify-content: center;
-        overflow: hidden;
-
-        canvas {
-            width: 100%;
-            height: 100%;
-        }
-    }
-}
