+++ apps/web/src/plugin/plugins/pie-chart/config.json
@@ -78,5 +78,36 @@
             ]
         }
     ],
-    "view": []
+    "view": [
+        {
+            "tag": "div",
+            "class": "ms-pie-chart",
+            "style": "display: flex; flex-direction: column; width: 100%; min-width: 100%; height: 100%; padding: 16px; background: var(--main-background);",
+            "children": [
+                {
+                    "tag": "Tooltip",
+                    "class": "ms-pie-chart__header",
+                    "style": "font-size: 16px; font-weight: bold;",
+                    "props": {
+                        "autoEllipsis": true
+                    },
+                    "params": ["title"]
+                },
+                {
+                    "tag": "div",
+                    "class": "ms-pie-chart__content",
+                    "style": "display: flex; flex: 1; align-items: center; justify-content: center; overflow: hidden;",
+                    "children": [
+                        {
+                            "tag": "PieChart",
+                            "style": "width: 100%; height: 100%; max-height: 40vh;",
+                            "adapter": {
+                                "model": "getAggregateHistory"
+                            }
+                        }
+                    ]
+                }
+            ]
+        }
+    ]
 }
