--- apps/web/src/plugin/plugins/pie-chart/configure/index.tsx
@@ -1,30 +1,0 @@
-import { forwardRef } from 'react';
-import { RenderConfig } from '../../../render';
-import type { ConfigureType, ViewConfigProps } from '../typings';
-
-interface ConfigPluginProps {
-    value: ViewConfigProps;
-    config: ConfigureType;
-    onOk: (data: any) => void;
-    onChange: (data: any) => void;
-}
-
-const Plugin = forwardRef((props: ConfigPluginProps, ref: any) => {
-    const { onOk, onChange, value, config } = props;
-
-    const handleSubmit = (data: any) => {
-        onOk(data);
-    };
-
-    return (
-        <RenderConfig
-            config={config}
-            onOk={handleSubmit}
-            ref={ref}
-            onChange={onChange}
-            value={value}
-        />
-    );
-});
-
-export default Plugin;
