--- apps/web/src/plugin/plugins/pie-chart/typings.d.ts
@@ -1,8 +1,0 @@
-export interface ViewConfigProps {
-    title: string;
-    entity: EntityOptionType;
-    time: number;
-    metrics: DataAggregateType;
-}
-
-export type ConfigureType = any;
