--- apps/web/src/plugin/plugins/pie-chart/view/style.less
@@ -1,28 +1,0 @@
-.ms-pie-chart {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    min-width: 100%;
-    height: 100%;
-    padding: @padding-md;
-    background: var(--main-background);
-
-    &__header {
-        font-size: @font-size-lg;
-        font-weight: @font-weight-bold;
-    }
-
-    &__content {
-        display: flex;
-        flex: 1;
-        align-items: center;
-        justify-content: center;
-        overflow: hidden;
-
-        canvas {
-            width: 100%;
-            height: 100%;
-            max-height: 40vh;
-        }
-    }
-}
