+++ apps/web/src/plugin/view-components/chart-view/components/basic-gauge-chart/gauge-controller/types.d.ts
