+++ apps/web/src/plugin/view-components/chart-view/index.tsx
@@ -3,3 +3,5 @@ export { default as BarChart } from './views/barChart';
 export { default as HorizonBarChart } from './views/horizonBarChart';
 export { default as LineChart } from './views/lineChart';
 export { default as RadarChart } from './views/radarChart';
+export { default as PieChart } from './views/pieChart';
+export { default as GaugeChart } from './views/gaugeChart';
