+++ apps/web/src/plugin/view-components/chart-view/views/barChart.tsx
@@ -5,26 +5,21 @@ import { getChartColor } from '@/plugin/utils';
 import BasicChart from '../components/basicChart';
 
 type BarChartOpts = ChartConfiguration<'bar', (string | number | null)[], string>;
-interface ChartDatasetsProps {
-    entityLabel: string;
-    entityValues: (string | number | null)[];
-}
 interface IProps {
-    chartDatasets: ChartDatasetsProps[];
-    chartLabels: string[];
+    chartDatasets: AdapterResult[];
     chartOptions?: BarChartOpts;
 }
-export default React.memo(({ chartDatasets = [], chartLabels = [], chartOptions }: IProps) => {
+export default React.memo(({ chartDatasets = [], chartOptions }: IProps) => {
     const customChartOptions = useMemo(() => {
         const chartColors = getChartColor(chartDatasets || []);
 
         const defaultOptions: BarChartOpts = {
             type: 'bar',
             data: {
-                labels: chartLabels,
-                datasets: (chartDatasets || []).map((chart: any, index: number) => ({
-                    label: chart.entityLabel,
-                    data: chart.entityValues,
+                labels: chartDatasets?.[0]?.data.map(item => item?.key),
+                datasets: (chartDatasets || []).map((chart, index) => ({
+                    label: chart?.entity?.label,
+                    data: (chart?.data || []).map(item => item?.value),
                     borderWidth: 1,
                     backgroundColor: chartColors[index],
                 })),
@@ -40,7 +35,7 @@ export default React.memo(({ chartDatasets = [], chartLabels = [], chartOptions 
             },
         };
         return merge(defaultOptions, chartOptions);
-    }, [chartDatasets, chartLabels, chartOptions]);
+    }, [chartDatasets, chartOptions]);
 
     return <BasicChart chartOptions={customChartOptions as ChartConfiguration} />;
 });
