+++ apps/web/src/plugin/view-components/index.ts
@@ -1,3 +1,11 @@
 export { default as Tooltip } from './tooltip';
 
-export { AreaChart, BarChart, HorizonBarChart, LineChart, RadarChart } from './chart-view';
+export {
+    AreaChart,
+    BarChart,
+    HorizonBarChart,
+    LineChart,
+    RadarChart,
+    PieChart,
+    GaugeChart,
+} from './chart-view';
